+++ brave/build.gradle
@@ -2,6 +2,6 @@ dependencies {
     api 'io.zipkin.brave:brave'
     api 'io.zipkin.brave:brave-instrumentation-http'
 
-    testImplementation project(':thrift0.13')
+    testImplementation project(':thrift0.14')
     testImplementation 'io.zipkin.brave:brave-instrumentation-http-tests'
 }
