+++ core/src/main/resources/com/linecorp/armeria/public_suffixes.txt
@@ -27,6 +27,7 @@
 *.customer-oci.com
 *.dapps.earth
 *.dev.adobeaemcloud.com
+*.devcdnaccesso.com
 *.dweb.link
 *.elb.amazonaws.com
 *.elb.amazonaws.com.cn
@@ -2498,6 +2499,7 @@ force.museum
 ford
 forde.no
 forex
+forgeblocks.com
 forgot.her.name
 forgot.his.name
 forli-cesena.it
@@ -2649,7 +2651,6 @@ fujisawa.iwate.jp
 fujisawa.kanagawa.jp
 fujishiro.ibaraki.jp
 fujitsu
-fujixerox
 fujiyoshida.yamanashi.jp
 fukagawa.hokkaido.jp
 fukaya.saitama.jp
@@ -6702,6 +6703,7 @@ rhcloud.com
 ri.it
 ri.us
 ribeirao.br
+ric.jelastic.vps-host.net
 rich
 richardli
 ricoh
@@ -8340,6 +8342,7 @@ voto
 voyage
 vpndns.net
 vpnplus.to
+vps-host.net
 vps.mcdir.ru
 vr.it
 vs.it
