+++ logback/build.gradle
@@ -1,5 +1,5 @@
 dependencies {
-    testImplementation project(':thrift0.13')
+    testImplementation project(':thrift0.14')
     // Logback
     api 'ch.qos.logback:logback-classic'
 }
