+++ thrift0.13/src/main/java/com/linecorp/armeria/common/thrift/ThriftJacksonSerializers.java
@@ -28,6 +28,7 @@ import org.apache.thrift.protocol.TMessage;
 import org.apache.thrift.protocol.TProtocol;
 import org.apache.thrift.protocol.TProtocolFactory;
 import org.apache.thrift.transport.TMemoryBuffer;
+import org.apache.thrift.transport.TTransportException;
 
 import com.fasterxml.jackson.core.JsonGenerator;
 import com.fasterxml.jackson.databind.BeanDescription;
@@ -37,6 +38,8 @@ import com.fasterxml.jackson.databind.SerializationConfig;
 import com.fasterxml.jackson.databind.SerializerProvider;
 import com.fasterxml.jackson.databind.ser.Serializers;
 import com.fasterxml.jackson.databind.ser.std.StdSerializer;
+
+import com.linecorp.armeria.common.util.Exceptions;
 
 final class ThriftJacksonSerializers extends Serializers.Base implements Serializable {
 
@@ -117,7 +120,15 @@ final class ThriftJacksonSerializers extends Serializers.Base implements Seriali
     }
 
     private static String serializeTBaseLike(Consumer<TProtocol> writer, boolean useNamedEnums) {
-        final TMemoryBuffer buffer = new TMemoryBuffer(1024);
+        final TMemoryBuffer buffer;
+        try {
+            buffer = new TMemoryBuffer(1024);
+        } catch (Exception ex) {
+            // TTransportException is added as a checked exception in Thrift 0.14.0
+            // However, a TTransportException should not be raised by the Armeria Thrift implementations.
+            return Exceptions.throwUnsafely(ex);
+        }
+
         final TProtocolFactory factory = useNamedEnums ? ThriftProtocolFactories.TEXT_NAMED_ENUM
                                                        : ThriftProtocolFactories.TEXT;
         final TProtocol protocol = factory.getProtocol(buffer);
