+++ thrift0.13/src/main/java/com/linecorp/armeria/internal/common/thrift/TByteBufTransport.java
@@ -16,93 +16,11 @@
 
 package com.linecorp.armeria.internal.common.thrift;
 
-import static java.util.Objects.requireNonNull;
-
-import javax.annotation.Nullable;
-
-import org.apache.thrift.transport.TTransport;
-import org.apache.thrift.transport.TTransportException;
-
 import io.netty.buffer.ByteBuf;
 
-public final class TByteBufTransport extends TTransport {
-
-    private final ByteBuf buf;
+public final class TByteBufTransport extends AbstractTByteBufTransport {
 
     public TByteBufTransport(ByteBuf buf) {
-        this.buf = requireNonNull(buf, "buf");
-    }
-
-    @Override
-    public void close() {}
-
-    @Override
-    public boolean isOpen() {
-        return true;
-    }
-
-    @Override
-    public void open() {}
-
-    @Override
-    public int read(byte[] buf, int off, int len) {
-        final int bytesRemaining = this.buf.readableBytes();
-        final int amtToRead = Math.min(len, bytesRemaining);
-        if (amtToRead > 0) {
-            this.buf.readBytes(buf, off, amtToRead);
-        }
-        return amtToRead;
-    }
-
-    @Override
-    public int readAll(byte[] buf, int off, int len) throws TTransportException {
-        final int bytesRemaining = this.buf.readableBytes();
-        if (len > bytesRemaining) {
-            throw new TTransportException("unexpected end of frame");
-        }
-
-        this.buf.readBytes(buf, off, len);
-        return len;
-    }
-
-    @Override
-    public void write(byte[] buf, int off, int len) {
-        this.buf.writeBytes(buf, off, len);
-    }
-
-    @Nullable
-    @Override
-    public byte[] getBuffer() {
-        final ByteBuf buf = this.buf;
-        if (!buf.hasArray())  {
-            return null;
-        } else {
-            return buf.array();
-        }
-    }
-
-    @Override
-    public int getBufferPosition() {
-        final ByteBuf buf = this.buf;
-        if (!buf.hasArray())  {
-            return 0;
-        } else {
-            return buf.arrayOffset() + buf.readerIndex();
-        }
-    }
-
-    @Override
-    public int getBytesRemainingInBuffer() {
-        final ByteBuf buf = this.buf;
-        if (buf.hasArray()) {
-            return buf.readableBytes();
-        } else {
-            return -1;
-        }
-    }
-
-    @Override
-    public void consumeBuffer(int len) {
-        buf.skipBytes(len);
+        super(buf);
     }
 }
